+++ example/basic/main.go
@@ -15,13 +15,13 @@ func main() {
 	ctx := context.Background()
 
 	// Configure OpenTelemetry with sensible defaults.
-	uptrace.ConfigureOpentelemetry(&uptrace.Config{
+	uptrace.ConfigureOpentelemetry(
 		// copy your project DSN here or use UPTRACE_DSN env var
-		DSN: "",
+		// uptrace.WithDSN("https://<key>@api.uptrace.dev/<project_id>"),
 
-		ServiceName:    "myservice",
-		ServiceVersion: "1.0.0",
-	})
+		uptrace.WithServiceName("myservice"),
+		uptrace.WithServiceVersion("1.0.0"),
+	)
 	// Send buffered spans and free resources.
 	defer uptrace.Shutdown(ctx)