+++ example/grpc/client/client.go
@@ -17,13 +17,15 @@ import (
 func main() {
 	ctx := context.Background()
 
-	uptrace.ConfigureOpentelemetry(&uptrace.Config{
+	// Configure OpenTelemetry with sensible defaults.
+	uptrace.ConfigureOpentelemetry(
 		// copy your project DSN here or use UPTRACE_DSN env var
-		DSN: "",
+		// uptrace.WithDSN("https://<key>@api.uptrace.dev/<project_id>"),
 
-		ServiceName:    "myservice",
-		ServiceVersion: "1.0.0",
-	})
+		uptrace.WithServiceName("myservice"),
+		uptrace.WithServiceVersion("1.0.0"),
+	)
+	// Send buffered spans and free resources.
 	defer uptrace.Shutdown(ctx)
 
 	target := os.Getenv("GRPC_TARGET")